+++ src/main/java/com/codetaylor/mc/pyrotech/modules/hunting/init/BlockInitializer.java
@@ -3,7 +3,9 @@ package com.codetaylor.mc.pyrotech.modules.hunting.init;
 import com.codetaylor.mc.athenaeum.registry.Registry;
 import com.codetaylor.mc.athenaeum.util.ModelRegistrationHelper;
 import com.codetaylor.mc.pyrotech.library.util.RegistryHelper;
+import com.codetaylor.mc.pyrotech.modules.core.block.spi.BlockBushBase;
 import com.codetaylor.mc.pyrotech.modules.hunting.ModuleHunting;
+import com.codetaylor.mc.pyrotech.modules.hunting.block.BlockButchersBlock;
 import com.codetaylor.mc.pyrotech.modules.hunting.block.BlockCarcass;
 import com.codetaylor.mc.pyrotech.modules.hunting.item.ItemBlockCarcass;
 import com.codetaylor.mc.pyrotech.modules.hunting.tile.TileCarcass;
@@ -19,6 +21,8 @@ public final class BlockInitializer {
       registry.registerBlock(block, new ItemBlockCarcass(block), BlockCarcass.NAME);
     }
 
+    registry.registerBlockWithItem(new BlockButchersBlock(), BlockButchersBlock.NAME);
+
     RegistryHelper.registerTileEntities(
         registry,
         TileCarcass.class
@@ -31,7 +35,8 @@ public final class BlockInitializer {
     registry.registerClientModelRegistrationStrategy(() -> {
 
       ModelRegistrationHelper.registerBlockItemModels(
-          ModuleHunting.Blocks.CARCASS
+          ModuleHunting.Blocks.CARCASS,
+          ModuleHunting.Blocks.BUTCHERS_BLOCK
       );
     });
   }