+++ src/main/java/com/codetaylor/mc/pyrotech/modules/tech/basic/plugin/top/provider/CompostBinProvider.java
@@ -47,6 +47,7 @@ public class CompostBinProvider
 
     if (tileEntity instanceof TileCompostBin) {
       this.probeInfo = probeInfo;
+      this.delegate.setPlayer(player);
       this.delegate.display((TileCompostBin) tileEntity);
       this.probeInfo = null;
     }